requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
cerebras-cloud-sdk==1.5.0
pydantic==2.5.0 